RESPONSIVE NAVBAR and i18n

+++ src/main/java/fr/formation/afpa/controller/IndexController.java
@@ -6,9 +6,11 @@ import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
 import javax.persistence.Persistence;
+import javax.servlet.http.HttpSession;
 
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -23,36 +25,36 @@ import fr.formation.afpa.domain.Employee;
 
 @Controller
 //@RequestMapping(path="/admin")
+//@RequestMapping(value="account")
 public class IndexController {
 	EntityManagerFactory entityManagerFactory = Persistence.createEntityManagerFactory("unitBd");
 	EntityManager entityManager = entityManagerFactory.createEntityManager();
 
 	EmployeeDaoJpa dao = new EmployeeDaoJpa();
+	boolean isAuthenticated = false;
 
 	// injecter service
 
-//	private boolean isAuthenticated() {
-//		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-//		if (authentication == null || AnonymousAuthenticationToken.class.
-//				isAssignableFrom(authentication.getClass())) {
-//			return false;
-//		}
-//		return authentication.isAuthenticated();
-//	}
+	//	private boolean isAuthenticated() {
+	//		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+	//		if (authentication == null || AnonymousAuthenticationToken.class.
+	//				isAssignableFrom(authentication.getClass())) {
+	//			return false;
+	//		}
+	//		return authentication.isAuthenticated();
+	//	}
 
 	@RequestMapping(path = "/employees", method = RequestMethod.GET)
-	public ModelAndView getEmployees() {
+	public ModelAndView getEmployees(HttpSession session) {
 		ModelAndView mv = new ModelAndView();
-		//if(isAuthenticated()) {
+		if (session.getAttribute("username") != null) {
 			List <Employee> liste = dao.findAll();
 			mv.addObject("liste", liste);
 			mv.setViewName("employees");
 			return mv;
-//		} else {
-//			mv.setViewName("redirect:/connexion");
-//			return mv;
-//		}
-
+		}
+		mv.setViewName("redirect:/connexion");
+		return mv;
 	}
 
 	@RequestMapping(path = "/", method = RequestMethod.GET)
@@ -61,14 +63,19 @@ public class IndexController {
 	}
 
 	@RequestMapping(path = "/updateEmployee", method = RequestMethod.GET)
-	public ModelAndView updateEmployee(@RequestParam Integer empId) {
+	public ModelAndView updateEmployee(@RequestParam Integer empId, HttpSession session) {
 		ModelAndView mv = new ModelAndView();
-		Employee e = dao.findById(empId);
-		mv.addObject("employee", e);
-		List <Employee> managers = dao.managersList();
-		mv.addObject("managers", managers);
-		mv.setViewName("updateEmployee");
-		return mv;
+		if (session.getAttribute("username") != null) {
+			Employee e = dao.findById(empId);
+			mv.addObject("employee", e);
+			List <Employee> managers = dao.managersList();
+			mv.addObject("managers", managers);
+			mv.setViewName("updateEmployee");
+			return mv;
+		} else {
+			mv.setViewName("redirect:/connexion");
+			return mv;
+		}
 	}
 
 	@RequestMapping(path = "/contact", method = RequestMethod.GET)
@@ -77,9 +84,13 @@ public class IndexController {
 	}
 
 	@RequestMapping(path = "/deleteEmployee", method = RequestMethod.GET)
-	public String delete(@RequestParam Integer empId) {
-		dao.deleteById(empId);
-		return "redirect:/employees";
+	public String delete(@RequestParam Integer empId, HttpSession session) {
+		if (session.getAttribute("username") != null) {
+			dao.deleteById(empId);
+			return "redirect:/employees";
+		} else {
+			return "redirect:/connexion";
+		}
 	}
 
 	@RequestMapping(path = "/about", method = RequestMethod.GET)
@@ -88,16 +99,20 @@ public class IndexController {
 	}
 
 	@RequestMapping(path = "/addEmployee", method = RequestMethod.POST)
-	public String add(@ModelAttribute Employee employee, @RequestParam("startDateString") @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) Date formdate, @RequestParam("superiorEmpId") String superiorEmpId) {
-		if (formdate == null) {
-			Date d = new Date();
-			employee.setStartDate(d);
-			employee.setChef(dao.findById(Integer.parseInt(superiorEmpId)));
+	public String add(@ModelAttribute Employee employee, @RequestParam("startDateString") @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) Date formdate, @RequestParam("superiorEmpId") String superiorEmpId, HttpSession session) {
+		if (session.getAttribute("username") != null) {
+			if (formdate == null) {
+				Date d = new Date();
+				employee.setStartDate(d);
+				employee.setChef(dao.findById(Integer.parseInt(superiorEmpId)));
+			} else {
+				employee.setStartDate(formdate);
+			}
+			dao.save(employee);
+			return "redirect:/employees";
 		} else {
-			employee.setStartDate(formdate);
+			return "redirect:/connexion";
 		}
-		dao.save(employee);
-		return "redirect:/employees";
 	}
 
 	@RequestMapping(path = "/updateEmployee", method = RequestMethod.POST)
@@ -119,49 +134,63 @@ public class IndexController {
 	}
 
 	@RequestMapping(path = "/addEmployee", method = RequestMethod.GET)
-	public ModelAndView addEmployee() {
+	public ModelAndView addEmployee(HttpSession session) {
 		ModelAndView mv = new ModelAndView();
-		List <Employee> managers = dao.managersList();
-		mv.addObject("managers", managers);
-		mv.setViewName("addEmployee");
-		return mv;
+		if (session.getAttribute("username") != null) {
+			List <Employee> managers = dao.managersList();
+			mv.addObject("managers", managers);
+			mv.setViewName("addEmployee");
+			return mv;
+		} else {
+			mv.setViewName("redirect:/connexion");
+			return mv;
+		}
 	}
 
 	@RequestMapping(path = "/managers", method = RequestMethod.GET)
-	public ModelAndView getManagers() {
+	public ModelAndView getManagers(HttpSession session) {
 		ModelAndView mv = new ModelAndView();
-		List <Employee> liste = dao.managersList();
-		mv.addObject("liste", liste);
-		mv.setViewName("managers");
-		return mv;
+		if (session.getAttribute("username") != null) {
+			List <Employee> liste = dao.managersList();
+			mv.addObject("liste", liste);
+			mv.setViewName("managers");
+			return mv;
+		} else {
+			mv.setViewName("redirect:/connexion");
+			return mv;
+		}
 	}
 
 	@RequestMapping(path = "/parametres", method = RequestMethod.GET)
-	public ModelAndView getParametres() {
+	public ModelAndView getParametres(HttpSession session) {
 		ModelAndView mv = new ModelAndView();
-		List <Employee> liste = dao.findAll();
-		mv.addObject("liste", liste);
-		mv.setViewName("parametres");
+		if (session.getAttribute("username") != null) {
+			List <Employee> liste = dao.findAll();
+			mv.addObject("liste", liste);
+			mv.setViewName("parametres");
+		} else {
+			mv.setViewName("redirect:/connexion");
+		}
 		return mv;
 	}
 
 	@RequestMapping(path = "/connexion", method = RequestMethod.POST)
-	public String connexionFormulaire(@RequestParam ("username") String name, @RequestParam ("password") String pwd) {
-		ModelAndView mv = new ModelAndView();
+	public String connexionFormulaire(@RequestParam ("username") String name, @RequestParam ("password") String pwd, ModelMap map, HttpSession session) {
 		String username = "root";
 		String password = "123456";
 		if(name.equals(username) && pwd.equals(password)) {
-
+			session.setAttribute("username", name);
 			return "redirect:/employees";
 		} else {
+			map.put("error", "<div class=\"alert alert-danger\" role=\"alert\">Identifiants incorrects.</div>");
 			return "connexion";
 		}
 	}
 
 	@RequestMapping(path = "/deconnexion", method = RequestMethod.GET)
-	public String deconnexion(SessionStatus status) {
-		status.setComplete();
-		return "index";	
+	public String deconnexion(HttpSession session) {
+		session.removeAttribute("username");
+		return "redirect:/";	
 	}
 
 }
